move helper functions for quantstudio structure to global utils
